Fix trait bound build error for Resolution: From
error[E0277]: the trait bound `Resolution: From<(u16, u16)>` is not satisfied
  --> src/decoder/stateless/h264/v4l2.rs:45:26
   |
45 |         let resolution = Resolution::from((
   |                          ^^^^^^^^^^ the trait `From<(u16, u16)>` is not implemented for `Resolution`
   |
   = help: the trait `From<(u32, u32)>` is implemented for `Resolution`
   = help: for that trait implementation, expected `u32`, found `u16`

+++ src/decoder/stateless/h264/v4l2.rs
@@ -43,8 +43,8 @@ impl StatelessH264DecoderBackend for V4l2StatelessDecoderBackend {
         let mb_unit = 16;
         let map_unit = 16;
         let resolution = Resolution::from((
-            (sps.pic_width_in_mbs_minus1 + 1) * mb_unit,
-            (sps.pic_height_in_map_units_minus1 + 1) * map_unit,
+            (sps.pic_width_in_mbs_minus1 + 1) as u32 * mb_unit,
+            (sps.pic_height_in_map_units_minus1 + 1) as u32 * map_unit,
         ));
         self.device.set_resolution(resolution);
         Ok(())
